+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Remove Windows 10 and dsvm image support from Guacamole. ([#4890](https://github.com/microsoft/AzureTRE/issues/4890))
 
 ENHANCEMENTS:
-* Allow new template properties to be specified during template upgrades. Remove Template properties that no longer exist.
+* Allow new template properties to be specified during template upgrades. Remove Template properties that no longer exist. ([#4783](https://github.com/microsoft/AzureTRE/pull/4783))
 * Specify default_outbound_access_enabled = false setting for all subnets ([#4757](https://github.com/microsoft/AzureTRE/pull/4757))
 * Pin all GitHub Actions workflow steps to full commit SHAs to prevent supply chain attacks plus update to latest releases ([#4886](https://github.com/microsoft/AzureTRE/pull/4886))
 * Add Windows Server 2025 image support to Guacamole. ([#4890](https://github.com/microsoft/AzureTRE/issues/4890))